refactor(config): core.config 패키지 분리 + PEP 562 lazy (pydantic_settings)
core/config.py (567 lines) → core/config/ 패키지:
- core/config/_settings.py (NEW) — Settings(BaseSettings) 클래스만 격리
- core/config/__init__.py — 상수 / TOML / dataclass / __getattr__ lazy

`from core.config import settings` / `Settings` 는 PEP 562 __getattr__ 로
첫 access 시점에만 _settings.py 를 import → pydantic_settings 풀 트리
(~150 ms cumulative, 144 modules) 가 cold-start 에서 빠짐. 상수만
import 하는 사이트 (예: ANTHROPIC_PRIMARY, CODEX_BASE_URL) 는 pydantic
회피.

측정:
- import core.config 단독 cold: 189 ms → 34 ms (-82 %)
- modules: 308 → 164
- pydantic_settings in sys.modules after import core.config: True → False
- settings 첫 access 시 정상 lazy 로드 + singleton 동작

cold-start 전체 회수는 단계 2 (top-level callsite 함수-로컬 이동) 에서.
이 PR 은 구조 분리 + lazy 토대.

검증:
- ruff: clean
- mypy: 332 source files clean
- pytest: 4330 passed / 0 regression
- E2E: A (68.4) unchanged
